backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "MCP Chatbot Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.2",
    "fastmcp>=2.12.4",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "mcp-use>=1.3.10",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.1",
    "requests>=2.31.0",
    "google-api-python-client>=2.108.0",
    "pytz>=2025.2",
]
